move table to a separate file
